Abhin||"add payment integration"

+++ src/Component/cart/Cart.jsx
@@ -10,27 +10,25 @@ import { useEffect } from "react";
 const Cart = () => {
   const context = useContext(MyContext);
   const [item, setItem] = useState([]);
-  const { name, url, price } = context.objectToPass;
+  const { name, url, price, id } = context.objectToPass;
   const navigate = useNavigate();
   useEffect(() => {
-   
     if (item.length > 0) {
-      updatefirestore(item)
-      
+      updatefirestore(item);
     }
   }, [item]);
-  if(item.length==0){
-  const storedItems = localStorage.getItem("cartItems");
-  if (storedItems) {
-    setItem(JSON.parse(storedItems));
+  if (item.length == 0) {
+    const storedItems = localStorage.getItem("cartItems");
+    if (storedItems) {
+      setItem(JSON.parse(storedItems));
+    }
   }
-}
-const pageChange=(item)=>{
-  item.length>0?localStorage.setItem("cartItems", JSON.stringify(item)):''
-  navigate("/cart", { state: { name, url, price } });
-
-
-}
+  const pageChange = (item) => {
+    item.length > 0
+      ? localStorage.setItem("cartItems", JSON.stringify(item))
+      : "";
+    navigate("/cart", { state: { name, url, price, id } });
+  };
 
   const clickHandler = () => {
     const newItem = {
@@ -39,12 +37,27 @@ const pageChange=(item)=>{
           name: name,
           price: price,
           url: url,
+          id: id,
         },
       },
     };
-   setItem((prev) => [...prev, newItem]);
+
+    const existingItem = item.find((e) => e.Cart.item.id === id);
+
+    if (!existingItem) {
+      // If the item doesn't exist, add it
+      alert("Item added");
+      setItem((prev) => [...prev, newItem]);
+    } else {
+      // If the item exists, increment the count
+      alert("Item already exists");
+      item.forEach((val) => {
+        val.count++;
+      }); // Increment the count property of the existing item
+      pageChange(item);
+    }
   };
-  pageChange(item)
+  pageChange(item);
 
   const updatefirestore = (item) => {
     const auth = getAuth();
@@ -52,8 +65,7 @@ const pageChange=(item)=>{
       if (user) {
         const uid = user.uid;
         const washingtonRef = doc(db, "user", uid);
-        await updateDoc(washingtonRef,{item});
-        
+        await updateDoc(washingtonRef, { item });
       }
     });
   };
